+++ Docker-requirements.txt
@@ -8,4 +8,3 @@ streamlit-pandas-profiling
 ydata-profiling
 pyarrow
 toml
-openpyxl
