--- docs/api-reference.md
@@ -1,20 +1,0 @@
-# API Reference
-
-<!-- Code in General -->
-
-:::src.new_weather_sourcing
-
-<!-- Prediction Pipeline -->
-
-<!-- Preprocessing --> 
-
-:::src.prediction_pipeline.pre_processing.impute_missing_parking_data
-:::src.prediction_pipeline.pre_processing.impute_missing_weather_values
-:::src.prediction_pipeline.pre_processing.join_sensor_weather_visitorcenter
-:::src.prediction_pipeline.pre_processing.visitor_center_processing_script
-:::src.prediction_pipeline.pre_processing.weather_data_sourcing_and_processing
-
-<!-- Sourcing Data -->
-
-:::src.prediction_pipeline.sourcing_data.source_historic_parking_data
-:::src.prediction_pipeline.sourcing_data.source_real_time_parking_data
