--- docs/executive-summary.md
@@ -1,1 +1,0 @@
-# Executive Summary
