--- docs/findings-and-results.md
@@ -1,1 +1,0 @@
-# Findings and Results
