--- docs/future-work.md
@@ -1,1 +1,0 @@
-# Recommendations and Future Work
