--- docs/index.md
@@ -1,10 +1,0 @@
-# bavarian forest project 
-
-This is the official documenatation for bavarian forest project.
-
-# Team Members:
-- Ayesha 
-- Anthony
-- Patricio
-- Manpa 
-
