--- docs/limitations-and-challenges.md
@@ -1,1 +1,0 @@
-# Limitations and Challenges
