--- docs/methodology-and-approach.md
@@ -1,1 +1,0 @@
-# Methodology and Approach
