--- docs/project-overview.md
@@ -1,1 +1,0 @@
-# Project Background and Context
