--- docs/references-and-appendix.md
@@ -1,1 +1,0 @@
-# References and Appendix
