--- requirements-mkdocs.txt
@@ -1,3 +1,0 @@
-mkdocs
-mkdocstrings[python]
-mkdocs-material
