+++ requirements.txt
@@ -1,11 +1,5 @@
 pandas
 jupyter
-openpyxl
-prettytable
-plotly
-pycaret
-pyarrow
-numpy
 plotly
 matplotlib
 chardet
